NSC-tx-lookup-service/requirements.txt
alembic>=1.13
confluent-kafka>=2.3
fastapi>=0.110
httpx>=0.26
PyJWT>=2.8
psycopg[binary]>=3.1
azure-monitor-opentelemetry>=1.6
sqlalchemy>=2.0
uvicorn[standard]>=0.27
